+++ src/dataconnect/types.ts
@@ -1,5 +1,3 @@
-import { FirebaseError } from "firebase-admin";
-
 // Schema is a singleton, so we always call it 'main'
 export const SCHEMA_ID = "main";
 
@@ -251,75 +249,3 @@ interface ImpersonationUnauthenticated {
 export type Impersonation = ImpersonationAuthenticated | ImpersonationUnauthenticated;
 
 /** End Dataplane Client Types */
-
-/** Start Gemini client types */
-
-export interface CallCloudAiCompanionRequest {
-  servicePath: string;
-  naturalLanguageQuery: string;
-  chatHistory: ChatMessage[];
-}
-
-export interface CloudAICompanionInput {
-  preamble?: string;
-  messages: CloudAICompanionMessage[];
-}
-
-export interface CloudAICompanionMessage {
-  content: string;
-  author: string;
-}
-
-export interface ExperienceContext {
-  experience?: string;
-  agent?: string;
-  task?: string;
-}
-export interface ClientContext {
-  name: string;
-  additionalContext: {
-    "@type": string;
-    fdcInfo: {
-      serviceId: string;
-      fdcServiceName: string;
-      requiresQuery: boolean;
-    };
-  };
-}
-
-export interface CloudAICompanionRequest {
-  messageId?: string;
-  topic?: string;
-  input: CloudAICompanionInput;
-
-  // product context -- required
-  experienceContext: ExperienceContext;
-
-  // Client context (e.g. IDE name, version, etc)
-  clientContext?: ClientContext;
-
-  // Unused
-  // // The GCP resources that the code generation process needs to reference
-  // backendResourcesContext?: BackendResourcesContext;
-
-  // // Additional user content not captured in the `input` field above
-  // inputDataContext?: InputDataContext;
-}
-
-export interface CloudAICompanionResponse {
-  output: {
-    messages: ChatMessage[];
-  };
-  error?: FirebaseError
-}
-
-export interface FdcRequestInfo {
-  serviceId: string;
-  fdcServiceName: string;
-  requiresQuery: boolean;
-}
-
-export interface ChatMessage {
-  content: string;
-  author: string;
-}
